refactor: restructure LID into bolna/lid/ with LIDBackend base class
