+++ bolna/transcriber/__init__.py
@@ -10,4 +10,4 @@ from .elevenlabs_transcriber import ElevenLabsTranscriber
 from .smallest_transcriber import SmallestTranscriber
 from .openai_transcriber import OpenAITranscriber
 from .transcriber_pool import TranscriberPool
-from .lid_provider import AzureLID, LIDProvider, SarvamLID
+from bolna.lid import AzureLID, LIDProvider, SarvamLID
